requirements.txt
# 实时EMG姿态预测系统依赖

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# 数据处理
numpy>=1.24.0
h5py>=3.8.0

# 配置管理
pyyaml>=6.0

# 可视化与日志
tensorboard>=2.12.0

# 其他工具
tqdm>=4.65.0

